backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.7.0
numpy==1.26.4
scikit-learn==1.5.0
httpx==0.25.0
redis==5.0.4
supabase==2.4.0
